🎯 Phase 1 continued fixes: CS8121 pattern matching, CS7036 method params, CS0173 type conversion, CS0201/CS0200/CS1501 - systematic elimination of critical errors ongoing

+++ src/BotCore/Services/TimeOptimizedStrategyManager.cs
@@ -237,7 +237,7 @@ namespace BotCore.Services
             var volume = CalculateVolumeProfile(bars);
             
             // Market microstructure features
-            var bidAskSpread = data.Bid > 0 && data.Ask > 0 ? (data.Ask - data.Bid) / data.Bid : 0.001m;
+            var bidAskSpread = data.Bid > 0 && data.Ask > 0 ? (decimal)((data.Ask - data.Bid) / data.Bid) : 0.001m;
             var imbalance = CalculateOrderBookImbalance(data);
             
             // Time-based features
